oneplus: import pn5xx nfc driver from OnePlusOSS/android_kernel_oneplus_msm8998
