+++ arch/arm/boot/dts/qcom/oneplus5-common.dtsi
@@ -671,24 +671,18 @@
 };
 
 &i2c_6 { /* BLSP1 QUP6 (NFC) */
-	status = "okay";
 	nq@28 {
-		/*
-		compatible = "qcom,nq-nci";
-		reg = <0x28>;
-		qcom,nq-irq = <&tlmm 92 0x00>;
-		qcom,nq-ven = <&tlmm 12 0x00>;
-		qcom,nq-firm = <&tlmm 93 0x00>;
-		qcom,nq-clkreq = <&pm8998_gpios 21 0x00>;
-		qcom,nq-esepwr = <&tlmm 116 0x00>;
-		*/
+		status = "disabled";
+	};
+
+	pn5xx@28 {
 		compatible = "nxp,pn544";
 		reg = <0x28>;
 		nxp,pn544-irq = <&tlmm 92 0x00>;
 		nxp,pn544-ven = <&tlmm 12 0x00>;
 		nxp,pn544-fw-dwnld = <&tlmm 93 0x00>;
 		nxp,pn544-clk-gpio = <&pm8998_gpios 21 0x00>;
-		nxp,pn544-ese-pwr = <&tlmm 37  0x00>;
+		nxp,pn544-ese-pwr = <&tlmm 37 0x00>;
 		nfc_voltage_s4-supply = <&pm8998_s4>;
 		nxp,pn544-wake-up = <&tlmm 116 0x00>;
 		interrupt-parent = <&tlmm>;
