+++ arch/arm64/configs/lineage_oneplus5_defconfig
@@ -1115,8 +1115,8 @@ CONFIG_NFC=y
 #
 # CONFIG_NFC_PN533 is not set
 # CONFIG_NFC_SIM is not set
-CONFIG_NFC_PN544_OEM=y
 CONFIG_NFC_NQ=y
+CONFIG_NFC_PN5XX=y
 # CONFIG_LWTUNNEL is not set
 CONFIG_IPC_ROUTER=y
 CONFIG_IPC_ROUTER_SECURITY=y
