+++ drivers/nfc/pn544/Makefile
@@ -8,4 +8,3 @@ pn544_mei-objs  = mei.o
 obj-$(CONFIG_NFC_PN544)     += pn544.o
 obj-$(CONFIG_NFC_PN544_I2C) += pn544_i2c.o
 obj-$(CONFIG_NFC_PN544_MEI) += pn544_mei.o
-obj-$(CONFIG_NFC_PN544_OEM) += pn544_oem.o
