--- drivers/nfc/pn544/pn544_oem.h
@@ -1,93 +1,0 @@
-/*
- * Copyright (C) 2010 Trusted Logic S.A.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- */
- /******************************************************************************
-  *
-  *  The original Work has been changed by NXP Semiconductors.
-  *
-  *  Copyright (C) 2015 NXP Semiconductors
-  *
-  *  Licensed under the Apache License, Version 2.0 (the "License");
-  *  you may not use this file except in compliance with the License.
-  *  You may obtain a copy of the License at
-  *
-  *  http://www.apache.org/licenses/LICENSE-2.0
-  *
-  *  Unless required by applicable law or agreed to in writing, software
-  *  distributed under the License is distributed on an "AS IS" BASIS,
-  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  *  See the License for the specific language governing permissions and
-  *  limitations under the License.
-  *
-  ******************************************************************************/
-
-#define PN544_MAGIC 0xE9
-
-/*
- * PN544 power control via ioctl
- * PN544_SET_PWR(0): power off
- * PN544_SET_PWR(1): power on
- * PN544_SET_PWR(2): reset and power on with firmware download enabled
- */
-#define PN544_SET_PWR    _IOW(PN544_MAGIC, 0x01, unsigned int)
-
-/*
- * SPI Request NFCC to enable p61 power, only in param
- * Only for SPI
- * level 1 = Enable power
- * level 0 = Disable power
- */
-#define P61_SET_SPI_PWR    _IOW(PN544_MAGIC, 0x02, unsigned int)
-
-/* SPI or DWP can call this ioctl to get the current
- * power state of P61
- *
-*/
-#define P61_GET_PWR_STATUS    _IOR(PN544_MAGIC, 0x03, unsigned int)
-
-/* DWP side this ioctl will be called
- * level 1 = Wired access is enabled/ongoing
- * level 0 = Wired access is disalbed/stopped
-*/
-#define P61_SET_WIRED_ACCESS _IOW(PN544_MAGIC, 0x04, unsigned int)
-
-/*
-  NFC Init will call the ioctl to register the PID with the i2c driver
-*/
-#define P544_SET_NFC_SERVICE_PID _IOW(PN544_MAGIC, 0x05, unsigned int)
-
-typedef enum p61_access_state{
-    P61_STATE_INVALID = 0x0000,
-    P61_STATE_IDLE = 0x0100, /* p61 is free to use */
-    P61_STATE_WIRED = 0x0200,  /* p61 is being accessed by DWP (NFCC)*/
-    P61_STATE_SPI = 0x0400, /* P61 is being accessed by SPI */
-    P61_STATE_DWNLD = 0x0800, /* NFCC fw download is in progress */
-    P61_STATE_SPI_PRIO = 0x1000, /*Start of p61 access by SPI on priority*/
-    P61_STATE_SPI_PRIO_END = 0x2000, /*End of p61 access by SPI on priority*/
-    P61_STATE_SPI_END = 0x4000,
-}p61_access_state_t;
-
-
-struct pn544_i2c_platform_data {
-    unsigned int irq_gpio;
-    unsigned int ven_gpio;
-    unsigned int firm_gpio;
-    unsigned int ese_pwr_gpio; /* gpio to give power to p61, only TEE should use this */
-    unsigned int clk_gpio;
-    unsigned int wake_up_gpio;
-    const char *clk_src_name;
-};
